requirements.txt
nltk
num2words
pydub
scipy
numpy
gitpython